[ADD] GET borrower requests

+++ global_models/borrow_request_response.go
@@ -8,8 +8,8 @@ type BorrowRequestResponse struct {
 	UserId          uint       `json:"user_id"`
 	Firstname       string     `json:"firstname"`
 	Lastname        string     `json:"lastname"`
-	RequestedAmount float64    `json:"requested_amount"`
-	RemainingAmount *float64   `json:"remaining_amount"`
+	RequestedAmount int64      `json:"requested_amount"`
+	RemainingAmount *int64     `json:"remaining_amount"`
 	RequestedAt     time.Time  `json:"requested_at"`
 	DueDate         *time.Time `json:"due_date"`
 }
